+++ pytket/extensions/aqt/backends/aqt.py
@@ -11,46 +11,56 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 import json
 import time
 from ast import literal_eval
-from typing import Dict, List, Optional, Sequence, Tuple, Union, cast, Any
-
+from typing import Any
+from typing import cast
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Sequence
+from typing import Tuple
+from typing import Union
 from requests import put
-from pytket.backends import Backend, CircuitStatus, ResultHandle, StatusEnum
+
+from pytket.backends import Backend
+from pytket.backends import CircuitStatus
+from pytket.backends import ResultHandle
+from pytket.backends import StatusEnum
 from pytket.backends.backend import KwargTypes
-from pytket.backends.backendinfo import BackendInfo, fully_connected_backendinfo
+from pytket.backends.backend_exceptions import CircuitNotRunError
+from pytket.backends.backendinfo import BackendInfo
+from pytket.backends.backendinfo import fully_connected_backendinfo
 from pytket.backends.backendresult import BackendResult
 from pytket.backends.resulthandle import _ResultIdTuple
-from pytket.backends.backend_exceptions import CircuitNotRunError
-from pytket.circuit import Circuit, Node, OpType, Qubit  # type: ignore
-from pytket.passes import (  # type: ignore
-    BasePass,
-    SequencePass,
-    SynthesiseTket,
-    FullPeepholeOptimise,
-    FlattenRegisters,
-    RebaseCustom,
-    EulerAngleReduction,
-    DecomposeBoxes,
-    RenameQubitsPass,
-    SimplifyInitial,
-    auto_rebase_pass,
-)
-from pytket.predicates import (  # type: ignore
-    GateSetPredicate,
-    MaxNQubitsPredicate,
-    NoClassicalControlPredicate,
-    NoFastFeedforwardPredicate,
-    NoMidMeasurePredicate,
-    NoSymbolsPredicate,
-    Predicate,
-)
+from pytket.circuit import Circuit
+from pytket.circuit import OpType
+from pytket.circuit import Qubit
+from pytket.passes import auto_rebase_pass
+from pytket.passes import BasePass
+from pytket.passes import DecomposeBoxes
+from pytket.passes import EulerAngleReduction
+from pytket.passes import FlattenRegisters
+from pytket.passes import FullPeepholeOptimise
+from pytket.passes import RenameQubitsPass
+from pytket.passes import SequencePass
+from pytket.passes import SimplifyInitial
+from pytket.passes import SynthesiseTket
+from pytket.predicates import GateSetPredicate
+from pytket.predicates import MaxNQubitsPredicate
+from pytket.predicates import NoClassicalControlPredicate
+from pytket.predicates import NoFastFeedforwardPredicate
+from pytket.predicates import NoMidMeasurePredicate
+from pytket.predicates import NoSymbolsPredicate
+from pytket.predicates import Predicate
 from pytket.utils import prepare_circuit
 from pytket.utils.outcomearray import OutcomeArray
-from .._metadata import __extension_version__
+
+
 from .config import AQTConfig
+
+from ..extension_version import __extension_version__
 
 AQT_URL_PREFIX = "https://gateway.aqt.eu/marmot/"
 
@@ -135,7 +145,7 @@ class AQTBackend(Backend):
 
         self._header = {"Ocp-Apim-Subscription-Key": access_token, "SDK": "pytket"}
         self._backend_info: Optional[BackendInfo] = None
-        self._qm: Dict[Qubit, Node] = {}
+        self._qm: Dict[Qubit, Qubit] = {}
         if device_name in _DEVICE_INFO:
             self._backend_info = fully_connected_backendinfo(
                 type(self).__name__,
@@ -145,7 +155,8 @@ class AQTBackend(Backend):
                 _GATE_SET,
             )
             self._qm = {
-                Qubit(i): node for i, node in enumerate(self._backend_info.nodes)
+                Qubit(i): cast(Qubit, node)
+                for i, node in enumerate(self._backend_info.nodes)
             }
         self._MACHINE_DEBUG = False
 
@@ -205,9 +216,6 @@ class AQTBackend(Backend):
                     FlattenRegisters(),
                     RenameQubitsPass(self._qm),
                     self.rebase_pass(),
-                    SimplifyInitial(
-                        allow_classical=False, create_all_qubits=True, xcirc=_xcirc
-                    ),
                     EulerAngleReduction(OpType.Ry, OpType.Rx),
                 ]
             )
@@ -219,9 +227,6 @@ class AQTBackend(Backend):
                     FlattenRegisters(),
                     RenameQubitsPass(self._qm),
                     self.rebase_pass(),
-                    SimplifyInitial(
-                        allow_classical=False, create_all_qubits=True, xcirc=_xcirc
-                    ),
                     EulerAngleReduction(OpType.Ry, OpType.Rx),
                 ]
             )
@@ -239,7 +244,13 @@ class AQTBackend(Backend):
     ) -> List[ResultHandle]:
         """
         See :py:meth:`pytket.backends.Backend.process_circuits`.
-        Supported kwargs: none.
+
+        Supported kwargs:
+        - `postprocess`: apply end-of-circuit simplifications and classical
+          postprocessing to improve fidelity of results (bool, default False)
+        - `simplify_initial`: apply the pytket ``SimplifyInitial`` pass to improve
+          fidelity of results assuming all qubits initialized to zero (bool, default
+          False)
         """
         circuits = list(circuits)
         n_shots_list = Backend._get_n_shots_as_list(
@@ -252,6 +263,7 @@ class AQTBackend(Backend):
             self._check_all_circuits(circuits)
 
         postprocess = kwargs.get("postprocess", False)
+        simplify_initial = kwargs.get("postprocess", False)
 
         handles = []
         for i, (c, n_shots) in enumerate(zip(circuits, n_shots_list)):
@@ -260,6 +272,10 @@ class AQTBackend(Backend):
                 ppcirc_rep = ppcirc.to_dict()
             else:
                 c0, ppcirc_rep = c, None
+            if simplify_initial:
+                SimplifyInitial(
+                    allow_classical=False, create_all_qubits=True, xcirc=_xcirc
+                ).apply(c0)
             (aqt_circ, measures) = _translate_aqt(c0)
             if self._MACHINE_DEBUG:
                 handles.append(
@@ -383,7 +399,7 @@ def _translate_aqt(circ: Circuit) -> Tuple[List[List], str]:
     return (gates, json.dumps(measures))
 
 
-def _aqt_rebase() -> RebaseCustom:
+def _aqt_rebase() -> BasePass:
     return auto_rebase_pass({OpType.XXPhase, OpType.Rx, OpType.Ry})