+++ pytket/extensions/aqt/multi_zone_architecture/macro_architecture_graph.py
@@ -1,7 +1,23 @@
-from dataclasses import dataclass, field
+# Copyright 2020-2023 Cambridge Quantum Computing
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+from dataclasses import field
 from typing import NewType
 
-from networkx import Graph, shortest_path  # type: ignore
+from networkx import Graph  # type: ignore
+from networkx import shortest_path
 
 from .architecture import MultiZoneArchitecture
 
@@ -38,7 +54,7 @@ class MultiZoneMacroArch:
         return path
 
 
-def empty_macro_arch_from_backend(
+def empty_macro_arch_from_architecture(
     architecture: MultiZoneArchitecture,
 ) -> MultiZoneMacroArch:
     zones = Graph()