+++ tests/multi_zone/multi_zone_backend_test.py
@@ -1,15 +1,29 @@
-import pytest
-from pytket.backends import ResultHandle
+# Copyright 2020-2023 Cambridge Quantum Computing
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
+import pytest
+from pytket import Circuit
+from pytket.backends import ResultHandle
+from pytket.extensions.aqt.backends.aqt_multi_zone import AQTMultiZoneBackend
 from pytket.extensions.aqt.backends.aqt_multi_zone import (
-    AQTMultiZoneBackend,
     get_aqt_json_syntax_for_compiled_circuit,
-)
-from pytket.extensions.aqt.multi_zone_architecture.named_architectures import (
-    four_zones_in_a_line,
 )
 from pytket.extensions.aqt.multi_zone_architecture.circuit.multizone_circuit import (
     MultiZoneCircuit,
+)
+from pytket.extensions.aqt.multi_zone_architecture.named_architectures import (
+    four_zones_in_a_line,
 )
 
 
@@ -29,13 +43,13 @@ def test_not_implemented_functionality_throws(backend: AQTMultiZoneBackend) -> N
     }
     circuit = MultiZoneCircuit(four_zones_in_a_line, initial_placement, 16)
     with pytest.raises(NotImplementedError):
-        backend.process_circuits([circuit])
+        backend.process_circuits([circuit])  # type: ignore
     with pytest.raises(NotImplementedError):
-        backend.process_circuit(circuit)
+        backend.process_circuit(circuit)  # type: ignore
     with pytest.raises(NotImplementedError):
-        backend.run_circuits([circuit])
+        backend.run_circuits([circuit])  # type: ignore
     with pytest.raises(NotImplementedError):
-        backend.run_circuit(circuit)
+        backend.run_circuit(circuit)  # type: ignore
     with pytest.raises(NotImplementedError):
         backend.circuit_status(ResultHandle())
     with pytest.raises(NotImplementedError):
@@ -58,7 +72,17 @@ def test_valid_circuit_compiles(backend: AQTMultiZoneBackend) -> None:
     circuit.move_qubit(0, 1)
     circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
     circuit.measure_all()
-    circuit = backend.get_compiled_circuit(circuit)
+    circuit = backend.compile_manually_routed_multi_zone_circuit(circuit)
+
+
+def test_circuit_compiles(backend: AQTMultiZoneBackend) -> None:
+    circuit = Circuit(8)
+    circuit.CX(0, 1).CX(2, 3).CX(4, 5).CX(6, 7)
+    circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
+    circuit.CX(0, 1).CX(2, 3).CX(4, 5).CX(6, 7)
+    circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
+    circuit.measure_all()
+    backend.compile_circuit_with_routing(circuit)
 
 
 def test_invalid_circuit_does_not_compile(backend: AQTMultiZoneBackend) -> None:
@@ -76,7 +100,7 @@ def test_invalid_circuit_does_not_compile(backend: AQTMultiZoneBackend) -> None:
     circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
     circuit.measure_all()
     with pytest.raises(Exception):
-        backend.get_compiled_circuit(circuit)
+        backend.compile_manually_routed_multi_zone_circuit(circuit)
 
 
 def test_try_get_aqt_syntax_on_uncompiled_circuit_raises(
@@ -107,7 +131,7 @@ def test_compiled_circuit_has_correct_syntax(backend: AQTMultiZoneBackend) -> No
     circuit.move_qubit(0, 1)
     circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
     circuit.measure_all()
-    circuit = backend.get_compiled_circuit(circuit)
+    circuit = backend.compile_manually_routed_multi_zone_circuit(circuit)
     aqt_operation_list = get_aqt_json_syntax_for_compiled_circuit(circuit)
 
     initialized_zones: list[int] = []
@@ -147,6 +171,70 @@ def test_compiled_circuit_has_correct_syntax(backend: AQTMultiZoneBackend) -> No
             assert _is_valid_zop(operation[1][1], initialized_zones)
         else:
             raise Exception(f"Detected invalid operation type: {operation[0]}")
+    assert initialized_zones == [zone for zone in initial_placement]
+    assert number_initialized_qubits == 8
+
+
+def test_automatically_routed_circuit_has_correct_syntax(
+    backend: AQTMultiZoneBackend,
+) -> None:
+    initial_placement = {0: [0, 1, 2, 3], 1: [4, 5, 6, 7]}
+    circuit = Circuit(8)
+    circuit.CX(0, 1).CX(2, 3).CX(4, 5).CX(6, 7)
+    circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
+    circuit.CX(0, 1).CX(2, 3).CX(4, 5).CX(6, 7)
+    circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
+    circuit.measure_all()
+    mz_circuit = backend.compile_circuit_with_routing(circuit, initial_placement)
+
+    n_shuttles = mz_circuit.get_n_shuttles()
+    n_pswaps = mz_circuit.get_n_pswaps()
+
+    aqt_operation_list = get_aqt_json_syntax_for_compiled_circuit(mz_circuit)
+
+    initialized_zones: list[int] = []
+    number_initialized_qubits: int = 0
+    aqt_shuttles = 0
+    aqt_pswaps = 0
+    for i, operation in enumerate(aqt_operation_list):
+        if i < 2:
+            assert operation[0] == "INIT"
+        else:
+            assert operation[0] != "INIT"
+        if operation[0] == "INIT":
+            initialized_zones.append(operation[1][0])
+            number_initialized_qubits += operation[1][1]
+        elif operation[0] in ["X", "Y", "Z"]:
+            assert len(operation) == 3
+            assert isinstance(operation[1], float)
+            assert len(operation[2]) == 1
+            assert _is_valid_zop(operation[2][0], initialized_zones)
+        elif operation[0] in ["MS"]:
+            assert len(operation) == 3
+            assert isinstance(operation[1], float)
+            assert len(operation[2]) == 2
+            assert _zop_addresses_in_same_zone(operation[2][0], operation[2][1])
+            assert _is_valid_zop(operation[2][0], initialized_zones)
+            assert _is_valid_zop(operation[2][1], initialized_zones)
+        elif operation[0] in ["SHUTTLE"]:
+            assert len(operation) == 3
+            assert isinstance(operation[1], int)
+            assert len(operation[2]) == 2
+            assert _zop_addresses_in_different_zones(operation[2][0], operation[2][1])
+            assert _is_valid_zop(operation[2][0], initialized_zones)
+            assert _is_valid_zop(operation[2][1], initialized_zones)
+            aqt_shuttles += 1
+        elif operation[0] in ["PSWAP"]:
+            assert len(operation) == 2
+            assert len(operation[1]) == 2
+            assert _zop_addresses_in_same_zone(operation[1][0], operation[1][1])
+            assert _is_valid_zop(operation[1][0], initialized_zones)
+            assert _is_valid_zop(operation[1][1], initialized_zones)
+            aqt_pswaps += 1
+        else:
+            raise Exception(f"Detected invalid operation type: {operation[0]}")
+    assert n_pswaps == aqt_pswaps
+    assert n_shuttles == aqt_shuttles
     assert initialized_zones == [zone for zone in initial_placement]
     assert number_initialized_qubits == 8