+++ tests/multi_zone/multi_zone_circuit_test.py
@@ -1,15 +1,37 @@
+# Copyright 2020-2023 Cambridge Quantum Computing
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import pytest
-
-from pytket.circuit import OpType  # type: ignore
-
+from pytket.circuit import Circuit, OpType
+from pytket.extensions.aqt.multi_zone_architecture.circuit.multizone_circuit import (
+    AcrossZoneOperationError,
+)
+from pytket.extensions.aqt.multi_zone_architecture.circuit.multizone_circuit import (
+    MoveError,
+)
 from pytket.extensions.aqt.multi_zone_architecture.circuit.multizone_circuit import (
     MultiZoneCircuit,
-    MoveError,
+)
+from pytket.extensions.aqt.multi_zone_architecture.circuit.multizone_circuit import (
     QubitPlacementError,
-    AcrossZoneOperationError,
 )
 from pytket.extensions.aqt.multi_zone_architecture.named_architectures import (
     four_zones_in_a_line,
+)
+
+from pytket.extensions.aqt.multi_zone_architecture.circuit_routing.route_zones import (
+    route_circuit,
 )
 
 
@@ -29,11 +51,20 @@ def circuit(initial_placement: dict[int, list[int]]) -> MultiZoneCircuit:
     return circuit
 
 
+@pytest.fixture()
+def circuit_precompile() -> Circuit:
+    circuit = Circuit(8)
+    circuit.CX(0, 1).CX(2, 3).CX(4, 5).CX(6, 7)
+    circuit.CX(1, 2).CX(3, 4).CX(5, 6).CX(7, 0)
+    circuit.measure_all()
+    return circuit
+
+
 def test_circuit_has_correct_init_gates_at_beginning(
     circuit: MultiZoneCircuit, initial_placement: dict[int, list[int]]
 ) -> None:
     circuit_placement = {}
-    for gate in circuit:
+    for gate in circuit.pytket_circuit:
         op = gate.op
         if "INIT" not in op.__str__():
             break
@@ -47,7 +78,7 @@ def test_circuit_contains_correct_number_of_moves_shuttles_swaps(
     circuit: MultiZoneCircuit,
 ) -> None:
     move_barriers, moves, shuttles, swaps = 0, 0, 0, 0
-    for gate in circuit:
+    for gate in circuit.pytket_circuit:
         op = gate.op
         if "MOVE_BARRIER" in op.__str__():
             move_barriers += 1
@@ -78,8 +109,6 @@ def test_move_on_missing_qubit_raises_placement_error(
 
 def test_add_barrier_throws_value_error(circuit: MultiZoneCircuit) -> None:
     with pytest.raises(ValueError):
-        circuit.add_barrier([0])
-    with pytest.raises(ValueError):
         circuit.add_gate(OpType.Barrier, [0])
 
 
@@ -95,3 +124,13 @@ def test_validation_of_circuit_with_operation_across_zones_throws(
 
 def test_validation_of_valid_circuit_does_not_throw(circuit: MultiZoneCircuit) -> None:
     circuit.validate()
+
+
+def test_circuit_routing(circuit_precompile: Circuit) -> None:
+    route_circuit(circuit_precompile, four_zones_in_a_line)
+
+
+def test_circuit_routing_with_initial_placement(
+    circuit_precompile: Circuit, initial_placement: dict[int, list[int]]
+) -> None:
+    route_circuit(circuit_precompile, four_zones_in_a_line, initial_placement)