+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 ~~~~~~~~~
 
+0.29.0 (October 2023)
+---------------------
+
+* Add simple automatic placement and routing for multi-zone architectures
+* Don't include ``SimplifyInitial`` in default passes; instead make it an option
+  to ``process_circuits()``.
+* Updated pytket version requirement to 1.21.
+
 0.28.0 (June 2023)
 ------------------